Updated to correct file ownership
I had to make this change in order to have all files in the /home/foundry/app directory owned by the user foundry.  They were previously owned by root.  This caused issues when trying to update 0.6.0 to 0.6.1 since the app user could not modify the files in /home/foundry/app